pyproject.toml
[tool.poetry]
name = "codehealth"
version = "0.1.0"
description = ""
authors = ["John Casey <jcasey@redhat.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.6"
GitPython = "^3.1.32"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
